server: Fixed an important bug when reassembling packets.
The packet reassembly context was a global variable and it was the same
for all the client connections. This created a situation where packets
were intermixed in the reassembly process and the resulting packet was
not at all what it was supposed to be.
This commit fix this issue by moving the packet reassembly context
inside the CList struct so packet reassembly is done in a per client
context.